[FD][APB-343] Avoid unnecessary creation of a MetricsController instance every time the metrics action is called
https://www.playframework.com/documentation/2.5.x/Migration24#routing

"The injected routes generator also supports the @ operator on routes, but it has a slightly different meaning (since everything is injected), if you prefix a controller with @, instead of that controller being directly injected, a JSR 330 Provider for that controller will be injected. This can be used, for example, to eliminate circular dependency issues, or if you want a new action instantiated per request."

Note that the controllers referenced from app.routes are not instantiated per request because they are annotated with @Singleton. Also attempting to remove the @ from them caused errors due to circular dependencies.