[FD][APB-431] Remove security checks from sandbox routes
The sandbox uses a test user that doesn't pass the security checks, so without this change the routes always returned a 401 when called in sandbox mode.

Because we aren't checking whether the user is a SA client or agent we can't use obtain the current user's identifiers any more and instead have to use hardcoded IDs when the response contains an identifier for the logged in user.